added user-detector-references component
