+++ AngularApp/projects/applens/src/app/modules/dashboard/update-detector-references/update-detector-references.component.html
@@ -1,0 +1,6 @@
+
+      <div class="detectorreferences-list" *ngIf = "detectorReferencesTable"> 
+        <fab-data-table [table]="detectorReferencesTable" [columnOptions]="" [allowColumnSearch]="true" [tableHeight]="'100%'" 
+        [descriptionColumnName] = "'Multiple'" [buttonText] = "'Update Selected'" (selectedItems) = updateDetectorReferences($event)>
+        </fab-data-table>
+        </div>
