+++ AngularApp/projects/diagnostic-data/src/lib/components/fab-data-table/fab-data-table.component.html
@@ -42,7 +42,7 @@
 </div>
 
 <div *ngIf = "buttonText != ''" style="display: flex;margin-bottom: -62px;margin-left: 90px;padding-top:20px;">
-  <fab-primary-button text="Update All" (onClick) = "returnSelectedItems()" [contentClass]="'mr-3'">
+  <fab-primary-button [text]="buttonText" (onClick) = "returnSelectedItems()" [contentClass]="'mr-3'">
   </fab-primary-button>
 </div>
